Changes after review
Change the kernel configuration to be a generic configuration so
other modules can use this parameter.
The parameter can be passed to build.sh and package.sh always, then
the script in the module decides to use it or not.
The variable build_dir name was not clear so it is changed to be
tmp_deb_dir which gives more descriptive.